+++ sample/Pizzeria.Store.StateManagement/Orders/Actions/AddPizzaToOrderFailureAction.cs
@@ -1,3 +1,3 @@
-namespace Pizzeria.Store.Contracts.Orders.Actions;
+namespace Pizzeria.Store.StateManagement.Orders.Actions;
 
 public record AddPizzaToOrderFailureAction(Guid PizzaId, string ErrorMessage);
