+++ sample/Pizzeria.Store.StateManagement/Orders/Actions/AddPizzaToOrderSuccessAction.cs
@@ -1,9 +1,9 @@
 using System.Diagnostics.CodeAnalysis;
 
-namespace Pizzeria.Store.Contracts.Orders.Actions;
+namespace Pizzeria.Store.StateManagement.Orders.Actions;
 
 [SuppressMessage(
     "Minor Code Smell",
     "S2094:Classes should not be empty",
     Justification = "False positive")]
-public record ClearOrderErrorAction;
+public record AddPizzaToOrderSuccessAction;
