+++ sample/Pizzeria.Store.StateManagement/Orders/Actions/StartOrderFailureAction.cs
@@ -1,6 +1,6 @@
-using Lewee.Contracts.StateManagement;
+using Lewee.StateManagement;
 
-namespace Pizzeria.Store.Contracts.Orders.Actions;
+namespace Pizzeria.Store.StateManagement.Orders.Actions;
 
 public record StartOrderFailureAction : IRequestErrorAction
 {
