+++ sample/Pizzeria.Store.StateManagement/Pizzas/Actions/LoadPizzasAction.cs
@@ -1,4 +1,4 @@
-namespace Pizzeria.Store.Contracts.Pizzas.Actions;
+namespace Pizzeria.Store.StateManagement.Pizzas.Actions;
 
 public record LoadPizzasAction
 {
