+++ sample/Pizzeria.Store.StateManagement/Pizzas/Actions/LoadPizzasFailureAction.cs
@@ -1,3 +1,3 @@
-namespace Pizzeria.Store.Contracts.Pizzas.Actions;
+namespace Pizzeria.Store.StateManagement.Pizzas.Actions;
 
 public record LoadPizzasFailureAction(string ErrorMessage);
