+++ sample/Pizzeria.Store.StateManagement/Pizzas/Actions/LoadPizzasSuccessAction.cs
@@ -1,5 +1,5 @@
 using Pizzeria.Store.Contracts.Pizzas;
 
-namespace Pizzeria.Store.Contracts.Pizzas.Actions;
+namespace Pizzeria.Store.StateManagement.Pizzas.Actions;
 
 public record LoadPizzasSuccessAction(IEnumerable<PizzaDto> Pizzas);
