+++ src/Lewee.Contracts/README.md
@@ -1,13 +1,10 @@
 # Lewee.Contracts
 
-Shared contracts for client-server communication and Fluxor state management in Lewee framework applications.
+Shared contracts for client-server communication in Lewee framework applications.
 
 ## Purpose
 
-This package provides the foundational contracts used for:
-
-- Message passing between servers and Blazor clients via SignalR
-- Interfaces for Fluxor state management actions
+This package provides the foundational contracts used for message passing between servers and Blazor clients via SignalR.
 
 ## Dependencies
 
@@ -46,35 +43,6 @@ await foreach (var message in channel.Reader.ReadAllAsync(cancellationToken))
 }
 ```
 
-### State Management Interfaces
-
-These interfaces define contracts for Fluxor actions used in client-side state management:
-
-| Interface | Description |
-|-----------|-------------|
-| `IRequestAction` | Base interface for actions that initiate a request, containing a `CorrelationId` |
-| `IRequestSuccessAction` | Interface for actions indicating successful request completion with `CorrelationId` |
-| `IQuerySuccessAction<T>` | Generic interface for successful query actions carrying `Data` of type `T` |
-| `IRequestErrorAction` | Interface for failed request actions with `CorrelationId` and `ErrorMessage` |
-| `IMessageReceivedAction` | Interface for actions dispatched when a server message is received |
-
-## Usage
-
-### Implementing State Management Actions
-
-```csharp
-// Request action to initiate a query
-public record GetOrdersAction(Guid CorrelationId) : IRequestAction;
-
-// Success action with query data
-public record GetOrdersSuccessAction(Guid CorrelationId, OrderDto[] Data)
-    : IQuerySuccessAction<OrderDto[]>;
-
-// Error action for failures
-public record GetOrdersErrorAction(Guid CorrelationId, string ErrorMessage)
-    : IRequestErrorAction;
-```
-
 ### Client-Side Message Deserialization
 
 On the client side (in Blazor applications), `ClientMessage` is deserialized back to the original type:
@@ -92,4 +60,4 @@ var messageBody = JsonSerializer.Deserialize(clientMessage.MessageJson, targetTy
 | `Lewee.Application` | `ClientEvent` uses `ClientMessage` as transport format for domain event notifications |
 | `Lewee.Infrastructure.AspNet` | `ClientEventHandler` sends `ClientMessage` objects via SignalR hubs |
 | `Lewee.Blazor` | `MessageDeserializer` processes `ClientMessage` objects and maps them to Fluxor actions |
-| `Lewee.StateManagement` | `ReducerExtensions` work with the state management interfaces for reducer patterns |
+| `Lewee.StateManagement` | References this package and provides state management action interfaces |
