+++ src/README.md
@@ -17,13 +17,13 @@ For development guidance in this directory, see the following instruction files:
 | `Lewee.Domain` | Domain layer abstractions including base classes for entities, aggregate roots, value objects, domain events, specifications, and repository interfaces |
 | `Lewee.Application` | Application layer with CQRS implementation using MediatR, including commands, queries, pipeline behaviors for logging, validation, and exception handling |
 | `Lewee.Shared` | Cross-cutting utilities including logging constants, HTTP context constants, request headers, and enum extension methods |
-| `Lewee.Contracts` | Shared contracts for client-server communication including client messages, event channels, and Fluxor state management action interfaces |
+| `Lewee.Contracts` | Shared contracts for client-server communication including client messages and event channels |
 | `Lewee.Infrastructure.Data` | Entity Framework Core integration with repository implementation, domain event dispatching, audit interceptors, and database configuration |
 | `Lewee.Infrastructure.PostgreSQL` | PostgreSQL-specific database configuration and setup using Npgsql with schema support and migration configuration |
 | `Lewee.Infrastructure.AspNet` | ASP.NET Core integration including authenticated user services, SignalR hub configuration, and correlation ID middleware |
 | `Lewee.Infrastructure.AspNet.WebApi` | FastEndpoints-based command and query endpoint base classes with MediatR integration and result handling |
 | `Lewee.Blazor` | Blazor integration with SignalR message receiving, HTTP correlation ID handling, and server health monitoring |
-| `Lewee.StateManagement` | Fluxor state management configuration and base classes for request/query state with reducer extension methods |
+| `Lewee.StateManagement` | Fluxor state management configuration and base classes for request/query state with reducer extension methods and action interfaces |
 | `Lewee.Playwright` | Playwright browser automation utilities for integration testing with page wrapper and extension methods |
 
 ## Key Requirements
