+++ tests/Lewee.Blazor.Tests.Web/Lewee.Blazor.Tests.Web.csproj
@@ -6,6 +6,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Lewee.Blazor\Lewee.Blazor.csproj" />
+    <ProjectReference Include="..\..\src\Lewee.Contracts\Lewee.Contracts.csproj" />
     <ProjectReference Include="..\..\src\Lewee.Infrastructure.AspNet\Lewee.Infrastructure.AspNet.csproj" />
     <ProjectReference Include="..\Lewee.Blazor.Tests.Contracts\Lewee.Blazor.Tests.Contracts.csproj" />
     <ProjectReference Include="..\Lewee.Tests.ServiceDefaults\Lewee.Tests.ServiceDefaults.csproj" />
