+++ tests/Lewee.Blazor.Tests.Web/MessageToActionMapper.cs
@@ -1,7 +1,7 @@
 using System.Diagnostics.CodeAnalysis;
 using Lewee.Blazor.Messaging;
 using Lewee.Blazor.Tests.Contracts;
-using Lewee.Contracts.StateManagement;
+using Lewee.StateManagement;
 
 namespace Lewee.Blazor.Tests.Web;
